// src/types/cliente.ts
export interface Cliente {
    id: string;
    numero_cliente: string;
    nome: string;
    documento: string;
    telefone: string;
    email: string;
    responsavel: string;
    tipo: string;
    origem: string;
    aniversario: string;
    cep: string;
    endereco: string;
    observacoes: string;
    senha: string;
    rua: string;
    numero: string;
    complemento: string;
    bairro: string;
    cidade: string;
    estado: string;
  }